Move util methods to Conditions also ErrorUpdateStatus
